refac: Move createFinding to finding.ts
